Fix bitmap tile plotter so tiled bitmaps continue out of the inital visible area.
svn path=/trunk/netsurf/; revision=5899

+++ amiga/bitmap.h
@@ -20,8 +20,8 @@
 #define AMIGA_BITMAP_H
 #include <exec/types.h>
 struct bitmap {
-	ULONG width;
-	ULONG height;
+	int width;
+	int height;
 	UBYTE *pixdata;
 };
